requirements.txt
cython
h5py
uwsgi
Django==1.11.15
social-auth-app-django
social-auth-core[saml]
python-social-auth
djangorestframework
markdown
django-filter
django-extensions
django-taggit
django-form-utils
django-crispy-forms
django-notifications-hq
django-taggit-templatetags
django-dirtyfields
psycopg2-binary
numexpr
shapely
Pillow
requests
requests-oauthlib
requests-toolbelt
celery[redis]<4
django-celery
django-chosen
opbeat
django-hstore==1.3.5
django-datatables-view
sregistry==0.0.96
django-gravatar2
pygments
google-api-python-client
google
oauth2client==3.0
retrying
django-rest-swagger
django-user-agents
django-guardian
python3-saml
